refactor: adding DTO responses and testing all endpoints

+++ ClubManager/ClubManager.Domain/DTOs/Financial/CreateEntityDTO.cs
@@ -4,8 +4,8 @@
     {
         public bool Internal { get; set; }
         public bool External { get; set; }
-        public int? ClubMemberId { get; set; }
-        public int? PlayerId { get; set; }
+        public long? ClubMemberId { get; set; }
+        public long? PlayerId { get; set; }
         public string EntityType { get; set; }
         public string EntityName { get; set; }
     }
