+++ ClubManager/ClubManager.Domain/DTOs/Financial/UpdateEntityDTO.cs
@@ -2,11 +2,11 @@
 {
     public class UpdateEntityDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public bool Internal { get; set; }
         public bool External { get; set; }
-        public int? ClubMemberId { get; set; }
-        public int? PlayerId { get; set; }
+        public long? ClubMemberId { get; set; }
+        public long? PlayerId { get; set; }
         public string EntityType { get; set; }
         public string EntityName { get; set; }
     }
