+++ ClubManager/ClubManager.Domain/DTOs/Infrastructures/CreateFacilityReservationDTO.cs
@@ -2,11 +2,11 @@
 {
     public class CreateFacilityReservationDTO
     {
-        public int FacilityId { get; set; }  
+        public long FacilityId { get; set; }  
         public DateTime StartTime { get; set; }  
         public DateTime EndTime { get; set; }  
         public string EventType { get; set; } 
         public string EventDescription { get; set; }  
-        public int ReservedUserId { get; set; }  
+        public long ReservedUserId { get; set; }  
     }
 }
