+++ ClubManager/ClubManager.Domain/DTOs/Infrastructures/CreateMaintenanceRequestDTO.cs
@@ -2,12 +2,12 @@
 {
     public class CreateMaintenanceRequestDTO
     {
-        public int FacilityId { get; set; }  
+        public long FacilityId { get; set; }  
         public string MaintenanceType { get; set; }  
         public string ProblemDescription { get; set; }  
         public string Priority { get; set; }  
         public DateTime RequestDate { get; set; }  
-        public int RequestedUserId { get; set; }  
+        public long RequestedUserId { get; set; }  
         public string Status { get; set; } 
     }
 }
