+++ ClubManager/ClubManager.Domain/DTOs/Infrastructures/UpdateFacilityDTO.cs
@@ -2,9 +2,9 @@
 {
     public class UpdateFacilityDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } 
-        public int FacilityCategoryId { get; set; }  
+        public long FacilityCategoryId { get; set; }  
         public string Location { get; set; }  
         public int Capacity { get; set; } 
         public string Description { get; set; }
