+++ ClubManager/ClubManager.Domain/DTOs/Infrastructures/UpdateFacilityReservationDTO.cs
@@ -2,12 +2,12 @@
 {
     public class UpdateFacilityReservationDTO
     {
-        public int Id { get; set; }
-        public int FacilityId { get; set; }  
+        public long Id { get; set; }
+        public long FacilityId { get; set; }  
         public DateTime StartTime { get; set; }  
         public DateTime EndTime { get; set; }  
         public string EventType { get; set; } 
         public string EventDescription { get; set; }  
-        public int ReservedUserId { get; set; }  
+        public long ReservedUserId { get; set; }  
     }
 }
