+++ ClubManager/ClubManager.Domain/DTOs/Infrastructures/UpdateMaintenanceRequestDTO.cs
@@ -2,13 +2,13 @@
 {
     public class UpdateMaintenanceRequestDTO
     {
-        public int Id { get; set; }
-        public int FacilityId { get; set; }  
+        public long Id { get; set; }
+        public long FacilityId { get; set; }  
         public string MaintenanceType { get; set; }  
         public string ProblemDescription { get; set; }  
         public string Priority { get; set; }  
         public DateTime RequestDate { get; set; }  
-        public int RequestedUserId { get; set; }  
+        public long RequestedUserId { get; set; }  
         public string Status { get; set; } 
     }
 }
