+++ ClubManager/ClubManager.Domain/DTOs/MembersTeams/UpdatePlayerContractDTO.cs
@@ -2,8 +2,8 @@
 {
     public class UpdatePlayerContractDTO
     {
-        public int? Id { get; set; }
-        public int PlayerId { get; set; }
+        public long? Id { get; set; }
+        public long PlayerId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public decimal Salary { get; set; }
