+++ ClubManager/ClubManager.Domain/DTOs/TrainingCompetition/CreateMatchStatisticDTO.cs
@@ -2,8 +2,8 @@
 {
     public class CreateMatchStatisticDTO
     {
-        public int MatchId { get; set; } 
-        public int PlayerId { get; set; } 
+        public long MatchId { get; set; } 
+        public long PlayerId { get; set; } 
         public int Goals { get; set; }
         public int Assists { get; set; }
         public int YellowCards { get; set; }
