+++ ClubManager/ClubManager.Domain/Entities/Infrastructures/Facility.cs
@@ -14,7 +14,7 @@
             Name = name;
         }
 
-        public void SetFacilityCategoryId(int facilityCategoryId)
+        public void SetFacilityCategoryId(long facilityCategoryId)
         {
             FacilityCategoryId = facilityCategoryId;
         }
